Fix bugs in adapt-one.sh

+++ adapt-one.sh
@@ -19,22 +19,27 @@ if [ -z "$python3_prints" ]; then
 else
     interpreter=python3
 fi
-echo $interpreter
+
+echo "Using $interpreter..."
+
 # parse the python file
-python3 lib/prepare.py "$file_path" 2>> log/prepare.log
+$interpreter lib/prepare.py "$file_path" 2>> log/prepare.log
 
-# Generate cases folder
-python3 lib/mutate.py "$output_dir" 2>> log/mutate.log
+if [ -e "$output_dir" ]; then
+    # Generate cases folder
+    $interpreter lib/mutate.py "$output_dir" 2>> log/mutate.log
 
-# Zip up cases folder
-cwd=$(pwd)
-cd "$output_dir"; zip -r cases.zip cases > /dev/null
-cd "$cwd"
+    # Zip up cases folder
+    cwd=$(pwd)
+    cd "$output_dir"; zip -r cases.zip cases > /dev/null
+    cd "$cwd"
 
-# Copy the description file
-description_path=$(dirname $file_path)/description.txt
-if [ -e "$description_path" ]; then
-    cp $description_path $output_dir
-else
-    touch $output_dir/description.txt
+    # Copy the description file
+    description_path=$(dirname $file_path)/description.txt
+
+    if [ -e "$description_path" ]; then
+        cp $description_path $output_dir
+    else
+        touch $output_dir/description.txt
+    fi
 fi
